+++ io/CMakeLists.txt
@@ -7,7 +7,7 @@ PCL_SUBSYS_OPTION(build "${SUBSYS_NAME}" "${SUBSYS_DESC}" ON)
 if(WIN32)
     PCL_SUBSYS_DEPEND(build "${SUBSYS_NAME}" DEPS ${SUBSYS_DEPS} OPT_DEPS openni openni2 ensenso davidSDK dssdk rssdk pcap png vtk)
 else(WIN32)
-    PCL_SUBSYS_DEPEND(build "${SUBSYS_NAME}" DEPS ${SUBSYS_DEPS} OPT_DEPS openni openni2 ensenso davidSDK dssdk pcap png vtk libusb-1.0)
+    PCL_SUBSYS_DEPEND(build "${SUBSYS_NAME}" DEPS ${SUBSYS_DEPS} OPT_DEPS openni openni2 ensenso davidSDK dssdk pcap png vtk libusb-1.0 librealsense)
 endif(WIN32)
 
 PCL_ADD_DOC("${SUBSYS_NAME}")
@@ -151,13 +151,23 @@ if(build)
     endif()
 
     if(WITH_RSSDK)
-        set(RSSDK_GRABBER_INCLUDES
+        set(REALSENSE_GRABBER_INCLUDES
             include/pcl/io/real_sense_grabber.h
         )
-        set(RSSDK_GRABBER_SOURCES
+        set(REALSENSE_GRABBER_SOURCES
             src/real_sense_grabber.cpp
-            src/real_sense/real_sense_device_manager.cpp
+            src/real_sense/sdk/real_sense_grabber.cpp
+            src/real_sense/sdk/real_sense_device_manager.cpp
         )
+    elseif(WITH_LIBREALSENSE)
+        set(REALSENSE_GRABBER_INCLUDES
+            include/pcl/io/real_sense_grabber.h
+        )
+        set(REALSENSE_GRABBER_SOURCES
+            src/real_sense_grabber.cpp
+            src/real_sense/librealsense/real_sense_grabber.cpp
+            src/real_sense/librealsense/real_sense_device_manager.cpp
+        )    
     endif()
 
     if(LIBUSB_1_FOUND)
@@ -226,7 +236,7 @@ if(build)
         ${ENSENSO_GRABBER_SOURCES}
         ${DAVIDSDK_GRABBER_SOURCES}
         ${DSSDK_GRABBER_SOURCES}
-        ${RSSDK_GRABBER_SOURCES}
+        ${REALSENSE_GRABBER_SOURCES}
         )
     if(PNG_FOUND)
       list(APPEND srcs
@@ -274,7 +284,7 @@ if(build)
         ${ENSENSO_GRABBER_INCLUDES}
         ${DAVIDSDK_GRABBER_INCLUDES}
         ${DSSDK_GRABBER_INCLUDES}
-        ${RSSDK_GRABBER_INCLUDES}
+        ${REALSENSE_GRABBER_INCLUDES}
         "include/pcl/${SUBSYS_NAME}/pxc_grabber.h"  # contains only depreciation note
         )
 
@@ -357,6 +367,8 @@ if(build)
 
     if(WITH_RSSDK)
       target_link_libraries(${LIB_NAME} ${RSSDK_LIBRARIES})
+    elseif(WITH_LIBREALSENSE)
+      target_link_libraries(${LIB_NAME} ${LIBREALSENSE_LIBRARIES})
     endif()
 
     if (PCAP_FOUND)
