+++ io/src/real_sense/librealsense/real_sense_device_manager.cpp
@@ -3,6 +3,7 @@
  *
  *  Point Cloud Library (PCL) - www.pointclouds.org
  *  Copyright (c) 2015-, Open Perception, Inc.
+ *  Copyright (c) 2016, Intel Corporation
  *
  *  All rights reserved.
  *
@@ -34,88 +35,15 @@
  *  POSSIBILITY OF SUCH DAMAGE.
  *
  */
-
 #include <pcl/io/io_exception.h>
-
-#include <pcl/io/real_sense/real_sense_device_manager.h>
+#include <pcl/io/real_sense/librealsense/real_sense_device_manager.h>
 
 boost::mutex pcl::io::real_sense::RealSenseDeviceManager::mutex_;
 
 using namespace pcl::io;
 
-/** Helper function to release a RealSense resource.
-    Useful as a deleter for a shared pointer holding RealSense resource. */
-template <typename T> void
-releasePXCResource (T* resource)
-{
-  if (resource)
-  {
-    resource->Release ();
-    resource = 0;
-  }
-}
-
-template <typename T> boost::shared_ptr<T>
-makePXCSharedPtr (T* resource)
-{
-  return boost::shared_ptr<T> (resource, releasePXCResource<T>);
-}
-
-boost::shared_ptr<PXCSession>
-createPXCSession ()
-{
-  PXCSession* s = PXCSession::CreateInstance ();
-  if (!s)
-    THROW_IO_EXCEPTION ("failed to create RealSense session");
-  return makePXCSharedPtr (s);
-}
-
-boost::shared_ptr<PXCCaptureManager>
-createPXCCaptureManager (PXCSession& session)
-{
-  PXCCaptureManager* cm = session.CreateCaptureManager ();
-  if (!cm)
-    THROW_IO_EXCEPTION ("failed to create RealSense capture manager");
-  return makePXCSharedPtr (cm);
-}
-
-boost::shared_ptr<PXCCapture>
-createPXCCapture (PXCSession& session, pxcUID iuid)
-{
-  PXCCapture* c;
-  if (session.CreateImpl (iuid, &c) < PXC_STATUS_NO_ERROR)
-    THROW_IO_EXCEPTION ("unable to create RealSense capture");
-  return makePXCSharedPtr (c);
-}
-
-boost::shared_ptr<PXCCapture::Device>
-createPXCCaptureDevice (PXCCapture& capture, pxcI32 didx)
-{
-  PXCCapture::Device* d;
-  d = capture.CreateDevice (didx);
-  if (!d)
-    THROW_IO_EXCEPTION ("unable to create RealSense capture device");
-  return makePXCSharedPtr (d);
-}
-
-/** Utility function to convert RealSense-style strings (which happen to
-  * consist of 2-byte chars) into standard library strings. */
-std::string
-toString (const pxcCHAR* pxc_string, size_t max_length)
-{
-  size_t i = 0;
-  while (i + 1 < max_length && pxc_string[i])
-    ++i;
-  std::string out (i + 1, '\0');
-  size_t j = 0;
-  while (j < i)
-    out[j] = pxc_string[j++];
-  return out;
-}
-
 pcl::io::real_sense::RealSenseDeviceManager::RealSenseDeviceManager ()
 {
-  session_ = createPXCSession ();
   populateDeviceList ();
 }
 
@@ -130,7 +58,7 @@ pcl::io::real_sense::RealSenseDeviceManager::captureDevice ()
   if (device_list_.size () == 0)
     THROW_IO_EXCEPTION ("no connected devices");
   for (size_t i = 0; i < device_list_.size (); ++i)
-    if (!device_list_[i].isCaptured ())
+    if (!device_list_[i].isCaptured)
       return (capture (device_list_[i]));
   THROW_IO_EXCEPTION ("all connected devices are captured by other grabbers");
   return (RealSenseDevice::Ptr ());  // never reached, needed just to silence -Wreturn-type warning
@@ -142,7 +70,7 @@ pcl::io::real_sense::RealSenseDeviceManager::captureDevice (size_t index)
   boost::mutex::scoped_lock lock (mutex_);
   if (index >= device_list_.size ())
     THROW_IO_EXCEPTION ("device with index %i is not connected", index + 1);
-  if (device_list_[index].isCaptured ())
+  if (device_list_[index].isCaptured)
     THROW_IO_EXCEPTION ("device with index %i is captured by another grabber", index + 1);
   return (capture (device_list_[index]));
 }
@@ -155,7 +83,7 @@ pcl::io::real_sense::RealSenseDeviceManager::captureDevice (const std::string& s
   {
     if (device_list_[i].serial == sn)
     {
-      if (device_list_[i].isCaptured ())
+      if (device_list_[i].isCaptured)
         THROW_IO_EXCEPTION ("device with serial number %s is captured by another grabber", sn.c_str ());
       return (capture (device_list_[i]));
     }
@@ -168,36 +96,12 @@ void
 pcl::io::real_sense::RealSenseDeviceManager::populateDeviceList ()
 {
   device_list_.clear ();
-
-  // Module description to match
-  PXCSession::ImplDesc module_desc = {};
-  module_desc.group = PXCSession::IMPL_GROUP_SENSOR;
-  module_desc.subgroup = PXCSession::IMPL_SUBGROUP_VIDEO_CAPTURE;
-
-  for (int m = 0;; m++)
+  for (int i = 0; i < context_.get_device_count (); i++)
   {
-    PXCSession::ImplDesc desc;
-    if (session_->QueryImpl (&module_desc, m, &desc) < PXC_STATUS_NO_ERROR)
-      break;
-    PXCCapture* capture;
-    if (session_->CreateImpl<PXCCapture> (&desc, &capture) < PXC_STATUS_NO_ERROR)
-      continue;
-    for (int j = 0;; j++)
-    {
-      PXCCapture::DeviceInfo device_info;
-      if (capture->QueryDeviceInfo (j, &device_info) < PXC_STATUS_NO_ERROR)
-        break;
-      if (device_info.streams & PXCCapture::STREAM_TYPE_DEPTH)
-      {
-        const size_t MAX_SERIAL_LENGTH = sizeof (device_info.serial) / sizeof (device_info.serial[0]);
-        std::string serial = toString (device_info.serial, MAX_SERIAL_LENGTH);
-        device_list_.push_back (DeviceInfo ());
-        device_list_.back ().serial = serial;
-        device_list_.back ().iuid = desc.iuid;
-        device_list_.back ().didx = j;
-      }
-    }
-    capture->Release ();
+    device_list_.push_back (DeviceInfo ());
+    device_list_.back ().serial = context_.get_device (i)->get_serial ();
+    device_list_.back ().index = i;
+    device_list_.back ().isCaptured = false;
   }
 }
 
@@ -206,17 +110,8 @@ pcl::io::real_sense::RealSenseDeviceManager::capture (DeviceInfo& device_info)
 {
   // This is called from public captureDevice() functions and should already be
   // under scoped lock
-  if (!device_info.device_ptr.expired ())
-  {
-    return device_info.device_ptr.lock ();
-  }
-  else
-  {
-    RealSenseDevice::Ptr device (new RealSenseDevice (device_info.serial));
-    device->capture_ = createPXCCapture (*session_, device_info.iuid);
-    device->device_ = createPXCCaptureDevice (*device->capture_, device_info.didx);
-    device_info.device_ptr = device;
-    return device;
-  }
+  RealSenseDevice::Ptr device (new RealSenseDevice (device_info.serial));
+  device->device_ = context_.get_device (device_info.index);
+  device_info.isCaptured = true;
+  return device;
 }
-
