+++ visualization/tools/CMakeLists.txt
@@ -60,5 +60,7 @@ endif()
 if(WITH_RSSDK)
     PCL_ADD_EXECUTABLE_OPT_BUNDLE(pcl_real_sense_viewer ${SUBSYS_NAME} real_sense_viewer.cpp)
     target_link_libraries(pcl_real_sense_viewer pcl_common pcl_io pcl_visualization)
+elseif(WITH_LIBREALSENSE)
+    PCL_ADD_EXECUTABLE_OPT_BUNDLE(pcl_real_sense_viewer ${SUBSYS_NAME} real_sense_viewer.cpp)
+    target_link_libraries(pcl_real_sense_viewer pcl_common pcl_io pcl_visualization)
 endif()
-
